refactor: 컴포넌트 이름 수정 Checkbox2 -> Checkbox
