+++ frontend/src/pages/DogProfile.tsx
@@ -1,26 +1,26 @@
 import { ResponseDogs, ResponseRecentMonthStatistics } from '@/api/dogs';
-import Topbar from '@/components/common/Topbar';
-import React, { ChangeEvent, useEffect, useState } from 'react';
+import { getUploadUrl } from '@/api/upload';
+import EditPhoto from '@/assets/icons/btn-edit-photo.svg';
+import CrownIcon from '@/assets/icons/ic-crown.svg';
 import Back from '@/assets/icons/ic-top-back.svg';
 import Edit from '@/assets/icons/ic-top-edit.svg';
-import CrownIcon from '@/assets/icons/ic-crown.svg';
-import Avatar from '@/components/common/Avatar';
-import { Divider } from '@/components/common/Divider';
-import DeleteDogModal from '@/components/DeleteDogModal';
-import BreedSearch from '@/components/BreedSearch';
-import { MaleIcon } from '@/components/icon/MaleIcon';
-import FemaleIcon from '@/components/icon/FemaleIcon';
-import { Checkbox } from '@/components/common/Checkbox2';
-import EditPhoto from '@/assets/icons/btn-edit-photo.svg';
-import { DogRegInfo } from '@/pages/Join';
-import ImageCropper from '@/components/ImageCropper';
-import { useCropStore } from '@/store/cropStore';
-import { maxLengthCheck } from '@/pages/JoinStep/DogDetailInfo';
-import { getUploadUrl } from '@/api/upload';
-import { dataURLtoFile } from '@/utils/dataUrlToFile';
-import { uploadImg, useDog } from '@/hooks/useDog';
-import { secondsToTimeFormat } from '@/utils/time';
 import Distance from '@/classes/Distance';
+import BreedSearch from '@/components/BreedSearch';
+import DeleteDogModal from '@/components/DeleteDogModal';
+import ImageCropper from '@/components/ImageCropper';
+import Avatar from '@/components/common/Avatar';
+import { Checkbox } from '@/components/common/Checkbox';
+import { Divider } from '@/components/common/Divider';
+import Topbar from '@/components/common/Topbar';
+import FemaleIcon from '@/components/icon/FemaleIcon';
+import { MaleIcon } from '@/components/icon/MaleIcon';
+import { uploadImg, useDog } from '@/hooks/useDog';
+import { DogRegInfo } from '@/pages/Join';
+import { maxLengthCheck } from '@/pages/JoinStep/DogDetailInfo';
+import { useCropStore } from '@/store/cropStore';
+import { dataURLtoFile } from '@/utils/dataUrlToFile';
+import { secondsToTimeFormat } from '@/utils/time';
+import { ChangeEvent, useEffect, useState } from 'react';
 interface Props {
     dog: ResponseDogs;
     statistics: ResponseRecentMonthStatistics;
