+++ frontend/src/pages/JoinStep/Agreements.tsx
@@ -1,4 +1,4 @@
-import { Checkbox } from '@/components/common/Checkbox2';
+import { Checkbox } from '@/components/common/Checkbox';
 import { Divider } from '@/components/common/Divider';
 
 interface AgreementsProps {
